refactor: Change placement of Categories component.
