+++ packages/map-template/src/components/Search/Categories/Categories.scss
